+++ Runtime/Auth/GamesAuthScope.cs
@@ -1,0 +1,11 @@
+// Copyright (c) BizSim Game Studios. All rights reserved.
+
+namespace BizSim.GPlay.Games
+{
+    public enum GamesAuthScope
+    {
+        Email,
+        Profile,
+        OpenId
+    }
+}